Adjustments for file extension and pdf convertion

+++ index.js
@@ -10,7 +10,7 @@ const BASE_DOWNLOAD_PATH = "https://i.nhentai.net/galleries/";
 const DOWNLOAD_BASE = "./downloads";
 
 async function process(nukeCode) {
-    
+
     const fullPageUrl = url.resolve(BASE_PATH, nukeCode);
 
     const sourceDir = path.join(DOWNLOAD_BASE, nukeCode);
@@ -21,13 +21,15 @@ async function process(nukeCode) {
 
     const fullDjName = DOC.querySelector("div#info>h1").innerHTML;
 
-    const fullPagesText = DOC.querySelector("div#info>div").innerHTML;
+    const pagesData = [...DOC.querySelectorAll("div.thumb-container>a>img")].map((img, index) => {
+        const imageExt = path.extname(img.attributes.getNamedItem("data-src").value);
+        const pageNumber = (index + 1);
+        return ({ imageExt, pageNumber });
+    })
 
     const coverSrc = DOC.querySelector("div#cover>a>img").attributes.getNamedItem("data-src").value;
 
     const curatedName = curateName(fullDjName);
-
-    const pages = parseInt(getNumericPart(fullPagesText), 10);
 
     const coverUrl = path.dirname(coverSrc);
 
@@ -35,21 +37,25 @@ async function process(nukeCode) {
 
     console.log(downloadNumber);
 
-    for (let i = 1; i <= pages; i++) {
-        const currentPage = `${i}.jpg`;
+    for (const index in pagesData) {
+        const { imageExt, pageNumber } = pagesData[index];
+        const currentPage = `${pageNumber}${imageExt}`;
         const downloadUrl = url.resolve(BASE_DOWNLOAD_PATH, path.join(downloadNumber, currentPage));
-        console.log(`Downloading image ${i} ${downloadUrl}`);
+        const parentDir = path.join(sourceDir, curatedName);
+        if (!fs.existsSync(parentDir)) {
+            console.log(`Mkdir directory ${parentDir}`);
+            fs.mkdirSync(parentDir, { recursive: true });
+        }
+        const imageFileName = paddCurrentPage(pageNumber) + imageExt;
+        const fileName = path.join(parentDir, imageFileName);
+        if (fs.existsSync(fileName)) {
+            continue;
+        }
+        console.log(`Downloading image ${pageNumber} ${downloadUrl}`);
         const response = await fetch(downloadUrl);
         if (response.status === 200) {
             const arrayBuffer = await response.arrayBuffer();
             const buffer = Buffer.from(arrayBuffer);
-            const parentDir = path.join(sourceDir, curatedName);
-            if (!fs.existsSync(parentDir)) {
-                console.log(`Mkdir directory ${parentDir}`);
-                fs.mkdirSync(parentDir, { recursive: true });
-            }
-            const imageFileName = paddCurrentPage(i) + ".jpg";
-            const fileName = path.join(parentDir, imageFileName);
             fs.writeFileSync(fileName, buffer);
             console.log(`Wrote ${fileName}`);
         }
@@ -60,7 +66,7 @@ async function process(nukeCode) {
 }
 
 function curateName(fullDjName = "") {
-    return fullDjName.replace(/( |)\[(([A-Z])\w+|(([A-Z])\w+( |)\w+))\]( |)/g, "");
+    return fullDjName.replace(/(( |)\[(([A-Z])\w+|(([A-Z])\w+( |)\w+))\]( |)|(\W )+|(.+\| )|(\[\W+\]))/g, "").replace(/(\.|\?|\\|\/)+/g, "");
 }
 
 function getNumericPart(fullPagesText = "") {
@@ -71,4 +77,4 @@ function paddCurrentPage(page = 0) {
     return ("0000" + page).substr(-4);
 }
 
-process("288285");
+process("264370");
